Support cmake variable BUILD_SHARED_LIBS
Use the cmake provided variable `BUILD_SHARED_LIBS` instead of the
custom `DDL_STATIC_LIBRARY` variable.

The variable `BUILD_SHARED_LIBS` is provided by cmake and respected by
the `add_library()` command to create the according shared of static
library as documented:

https://cmake.org/cmake/help/latest/variable/BUILD_SHARED_LIBS.html

Keep the `DDL_STATIC_LIBRARY` variable for backwards compatiblity. The
option value is used to provide a default value for the
`BUILD_SHARED_LIBS` variable. This means the variable works for the
configuration of a fresh build folder, but won't be respected when
changed afterwards. For this the `BUILD_SHARED_LIBS` variable should be
used.